// src/components/custom-popover/styles.tsx
import { alpha, styled } from "@mui/material/styles";

import { stylesMode } from "@/theme/styles";

import type { PopoverArrow } from "./types";

export const StyledArrow = styled("span", {
  shouldForwardProp: (prop) =>
    prop !== "size" && prop !== "placement" && prop !== "offset",
})<PopoverArrow>(({ placement, offset = 0, size = 0, theme }) => {
  const POSITION = -(size / 2) + 0.5;

  const alignmentStyles = {
    top: { top: POSITION, transform: "rotate(135deg)" },
    bottom: { bottom: POSITION, transform: "rotate(-45deg)" },
    left: { left: POSITION, transform: "rotate(45deg)" },
    right: { right: POSITION, transform: "rotate(-135deg)" },
    hCenter: { left: 0, right: 0, margin: "auto" },
    vCenter: { top: 0, bottom: 0, margin: "auto" },
  };

  return {
    width: size,
    height: size,
    position: "absolute",
    backdropFilter: "6px",
    borderBottomLeftRadius: size / 4,
    clipPath: "polygon(0% 0%, 100% 100%, 0% 100%)",
    backgroundColor: theme.palette.background.paper,
    border: `solid 1px ${alpha(theme.palette.grey["500"], 0.12)}`,
    [stylesMode.dark]: {
      border: `solid 1px ${alpha(theme.palette.common.black, 0.12)}`,
    },
    /**
     * Top
     */
    ...(placement === "top-left" && {
      ...alignmentStyles.top,
      left: offset,
    }),
    ...(placement === "top-center" && {
      ...alignmentStyles.top,
      ...alignmentStyles.hCenter,
    }),
    ...(placement === "top-right" && {
      ...alignmentStyles.top,
      right: offset,
    }),
    /**
     * Bottom
     */
    ...(placement === "bottom-left" && {
      ...alignmentStyles.bottom,
      left: offset,
    }),
    ...(placement === "bottom-center" && {
      ...alignmentStyles.bottom,
      ...alignmentStyles.hCenter,
    }),
    ...(placement === "bottom-right" && {
      ...alignmentStyles.bottom,
      right: offset,
    }),
    /**
     * Left
     */
    ...(placement === "left-top" && {
      ...alignmentStyles.left,
      top: offset,
    }),
    ...(placement === "left-center" && {
      ...alignmentStyles.left,
      ...alignmentStyles.vCenter,
    }),
    ...(placement === "left-bottom" && {
      ...alignmentStyles.left,
      bottom: offset,
    }),
    /**
     * Right
     */
    ...(placement === "right-top" && {
      ...alignmentStyles.right,
      top: offset,
    }),
    ...(placement === "right-center" && {
      ...alignmentStyles.right,
      ...alignmentStyles.vCenter,
    }),
    ...(placement === "right-bottom" && {
      ...alignmentStyles.right,
      bottom: offset,
    }),
  };
});
